changes asked by science soc
