+++ app/javascript/scavenger_hunt/scenes/HomePage/HomePage.scss
@@ -13,6 +13,7 @@
 .section {
   @extend .main;
   margin-top: 0px;
+  width: 60%;
 }
 
 .title {
