+++ app/javascript/scavenger_hunt/scenes/Questions/QuestionEleven/QuestionEleven.scss
@@ -1,5 +1,4 @@
 .letter-box {
-  width: 100%;
   border: 1px solid green !important;
   border-radius: 20px;
   font-size: 20px;
@@ -8,11 +7,6 @@
   margin-top: 20px;
 }
 
-.table {
-    width: 80%;
-    height: 300px;
-}
-
-.seq {
-  font-size: 12px;;
+.red {
+  color: 'red !important'
 }
