+++ app/javascript/scavenger_hunt/scenes/Questions/QuestionTwelve/QuestionTwelve.jsx
@@ -11,15 +11,15 @@ import './QuestionTwelve.scss';
 import { useGetUserAnswerQuery } from '../../../data/queries';
 import { useSaveUserAnswerMutation } from '../../../data/mutations';
 
-const  QuestionTwelve = ({ progress, setActiveStep, completed, setCompleted  }) => {
-  
+const QuestionTwelve = ({ progress, setActiveStep, completed, setCompleted  }) => {
+
   const [answer, setAnswer] = useState("");
   const [loading, setLoading ] = useState(false);
   const [toggle, setToggle] = useState(false);
   const [submitted, setSubmitted] = useState(false);
 
   const ans = Clues[11].answers[0];
-
+ 
   const { data: getUserAnswerQueryData, loading: getUserAnswerQueryLoading } = useGetUserAnswerQuery({
     variables: {
       question_number: 12
@@ -27,7 +27,7 @@ const  QuestionTwelve = ({ progress, setActiveStep, completed, setCompleted  }) 
   });
 
   const [saveUserAnswer, { loading: mutationLoading }] = useSaveUserAnswerMutation();
-
+  
   useEffect(() => {
     if(!getUserAnswerQueryLoading) {
       let persisted_user_answer = getUserAnswerQueryData.currentUser.answerTo;
@@ -45,12 +45,11 @@ const  QuestionTwelve = ({ progress, setActiveStep, completed, setCompleted  }) 
     setCompleted(newCompleted);
   }
 
-
   const handleSubmit = (event) => {
     event.preventDefault();
     setSubmitted(true);
     setLoading(true);
-    console.log("answer",  answer, ans )
+    console.log("progress",  progress)
     if (check(answer, ans)) {
       if(!mutationLoading){
         saveUserAnswer({
@@ -77,9 +76,7 @@ const  QuestionTwelve = ({ progress, setActiveStep, completed, setCompleted  }) 
     setAnswer(event.target.value);
   }
 
-  const Red = ({l}) => (<span style={{color : 'red'}}>{l}</span>)
-
-  return (
+  return(
     <Card>
       <Card.Header>
         <h1>Puzzle #12</h1>
@@ -87,56 +84,102 @@ const  QuestionTwelve = ({ progress, setActiveStep, completed, setCompleted  }) 
       <Card.Body>
         <form onSubmit={handleSubmit} >
           <div className="letter-box">
-            IODINE TENNESSINE 1000! CARBON HELIUM <Red l={'C'}/> POTASSIUM BORON LANTHANIUM CARBON POTASSIUM BORON OXYGEN ARGON <Red l={'D'}/>! <Red l={'M'}/> ARGON POTASSIUM SULFUR <Red l={'A'}/> RHENIUM URANIUM PHOSPHORUS FLUORINE OXYGEN <Red l={'R M'}/> ASTATINE HYDROGEN 11010111000!
-
-            <br/>
-            <br/>
-            Huh, what course was that again? 
-            <br/>
-            <br/>
-            (Please provide the answer in a 4-letter 4-number code, such as: AAAA 0000)
-
-          </div>
-          <div className="center-text">
-            <TextField required 
-              disabled={completed[progress].isCompleted}
-              id="question" 
-              label="Answer" 
-              variant="outlined"
-              aria-describedby="Write your answer here" 
-              value={answer} 
-              onChange={handleChange}
-            />
-          </div>
-          <Grid container justify="center" alignItems="center">
-          {
-            (completed[progress].isCompleted || (submitted && toggle)) &&
-              <CheckCircleOutlineIcon style={{ color: 'green', width: 50, height: 50}}/>
-          }
-          {
-            submitted && !toggle &&
-              <HighlightOffIcon style={{ color: 'red', width: 50, height: 50}}/>
-          }
-                 
-          </Grid>
-          {
-           (!toggle && !completed[progress].isCompleted)  &&
-            <div className="center-text">
-              <Button 
-                variant="primary" 
-                type="submit"
-                disabled={loading}
-              >
-                Submit
-              </Button>
+            <div>
+              Perhaps your name will join the ranks of these famous scientists someday after you graduate. <strong> That year seems far in the future, but it’ll come before you know it!</strong>
             </div>
-          }
+            <br/>
+            <br/>
+            <Grid container justify="center" direction="row" >
+              <Grid container item justify="flex-start"  xs={12} sm={6}>
+                1. Ada Lovelace
+              </Grid>
+              <Grid container item justify="flex-start"  xs={12} sm={6}>
+                2. Rosalind Franklin
+              </Grid>
+              <Grid container item justify="flex-start"  xs={12} sm={6}>
+                3. Mary Anning
+              </Grid>
+              <Grid container item justify="flex-start"  xs={12} sm={6}>
+                4. Jennifer Doudna
+              </Grid>
+              <Grid container item justify="flex-start"  xs={12} sm={6}>
+                5. Marie Curie
+              </Grid>
+              <Grid container item justify="flex-start"  xs={12} sm={6}>
+                6. Jane Goodall
+              </Grid>
+              <Grid container item justify="flex-start"  xs={12} sm={6}>
+                7. Katherine Johnson
+              </Grid>
+              <Grid container item justify="flex-start"  xs={12} sm={6}>
+                8. Mae C. Jemison
+              </Grid>
+              <Grid container item justify="flex-start"  xs={12} sm={6}>
+                9. Shirley Ann Jackson
+              </Grid>
+              <Grid container item justify="flex-start"  xs={12} sm={6}>
+                10. Gladys West
+              </Grid>
+
+            </Grid>
+    
+              <br/>
+              <br/>
+              Multiply the lifespans of scientists 1-3. This is value X.
+              <br/>
+              <br/>
+              Sum the birth years of scientists 4-6. This is value Y.
+              <br/>
+              <br/>
+              Sum the years that scientists 7-10 died, and divide it by 5. If the scientist is still alive, don’t include them in the calculation. This is value Z.
+              <br/>
+              <br/>
+              (X+Y)*Z / 101 - 271452 = ?
+              <br/>
+              <br/>
+            </div>
+
+            <div className="center-text">
+              <TextField required 
+                id="question" 
+                disabled={completed[progress].isCompleted}
+                label="Answer" 
+                variant="outlined"
+                aria-describedby="Write your answer here" 
+                value={answer} 
+                onChange={(e) => setAnswer(e.target.value)}
+              />
+            </div>
+          
+            <Grid container justify="center" alignItems="center">
+            {
+              (completed[progress].isCompleted || (submitted && toggle)) &&
+                <CheckCircleOutlineIcon style={{ color: 'green', width: 50, height: 50}}/>
+            }
+            {
+              submitted && !toggle &&
+                <HighlightOffIcon style={{ color: 'red', width: 50, height: 50}}/>
+            }
+                  
+            </Grid>
+            {
+            (!toggle && !completed[progress].isCompleted)  &&
+              <div className="center-text">
+                <Button 
+                  variant="primary" 
+                  type="submit"
+                  disabled={loading}
+                >
+                  Submit
+                </Button>
+              </div>
+            }
 
         </form>
       </Card.Body>
     </Card>
-  );
-};
 
+  )
+}
 
 export { QuestionTwelve };
