+++ app/javascript/scavenger_hunt/scenes/Questions/QuestionTwelve/QuestionTwelve.scss
@@ -1,4 +1,5 @@
 .letter-box {
+  width: 100%;
   border: 1px solid green !important;
   border-radius: 20px;
   font-size: 20px;
@@ -7,6 +8,11 @@
   margin-top: 20px;
 }
 
-.red {
-  color: 'red !important'
+.table {
+    width: 80%;
+    height: 300px;
+}
+
+.seq {
+  font-size: 12px;;
 }
