+++ app/javascript/scavenger_hunt/scenes/Questions/QuestionTwelve/index.js
@@ -1,1 +1,1 @@
-export { QuestionTwelve  } from './QuestionTwelve';
+export { QuestionTwelve } from './QuestionTwelve';
